Moved HttpManagementPayload to a seperate class
